delete sys.path[0]
When .setup() is called the first time, it now deletes sys.path[DELETE_SYS_PATH_INDEX], which by default is 0. Set it to None to not delete anything. The `ins_to` argument must be set to 'sys.path' (or sys.path) [which by default it is], in order for it to be executed.